Refactoring: Split files. Split functions
